[Client|Journal] move menu comps to folder
