+++ client/src/Components/Journal/JournalMenu/MenuEntry.tsx
